Fix SegmentStatusChecker exception (#3704)
SegmentStatusChecker is instantiated in the ControllerStarter constructor, right after PinotHelixResourceManager is instantiated. The SegmentStatusChecker (and all other periodic tasks) receive the _pinotHelixResourceManager object as a parameter in their constructor. However, the _pinotHelixResourceManager is not supposed to be used until PinotHelixResourceManager::start() is invoked. Only after invoking start() does the _pinotHelixResourceManager properly get initialized i.e. the _helixAdmin, _helixClusterName, _propertyStore get set. Until start() is invoked, the _helixAdmin, _helixClusterName and _propertyStore will be null in the _pinotHelixResourceManager.
As a result, even though we pass _pinotHelixResourceManager to all periodic tasks in their constructor, we cannot be accessing these objects, until after the invocation of start().