Second pass at my Apache doc items
Changes include:

1. Breaking out the table config section from the "Running Pinot in
Production" section. I think this makes the doc easier to follow for
table configs. I have a section here which includes a sample offline and
realtime table config. I am not sure if there are some configs included
as a way for users to run a local quick start. If there is I can link to
those files in the repo, otherwise I can include the full configs on
this page for users.

2. Adding links to metrics which I am able to find in the code. I need
to collect feedback on what to do with the others - Remove or I need
pointers to where I can find them in the code.