Updating metrics with names and URLs
The is two items I need before this is completed

1. I notice that the controller metrics appear to be generated dynamically
due to table names being present in most of the metrics. I need some advice
from the devs on how I should note these metric names so I can link to them
in the code. I want to make sure it is clear to the users.

2. I want to add sample table configs in the "in_production.rst"
section. I feel this would be best to be added in the git repo since
they will be useful for the users in the quick start guides. I need Jack
to tell me where in the repo I should put these files. Once they are in
the repo I will link to them from that section.