Added a percent threshold to consider startup of services
In installations where each server may host 1000s of tabls, and 10s of
thousands of segments, it is useful in practice to declare a server
as STARTED when "most" segments of "most" of the tables are ready (i.e.
externalview and idealstate match, as do currentstate and idealstate).

This PR makes an attempt at defining "most" (on an experimental basis).

For now, we disregard the number of (helix) partitions within each resource.
Instead, we count the number of resources that have converged state
and declare success if a (configured) percent are ready.

This is still experimental, and it is suggested to set the percent
to 100 all the time (or, not set it at all, and let the default ride).