ReadTheDocs documentation for Table Configs, Monitoring, and Deployment (#3975)
* First pass of 'Running Pinot in Production' docs

First pass of 'Running Pinot in Production' docs, which includes
sections on

* Deploying Pinot and the ordering
* Table configs and what each setting does
* Monitoring (still needs proper metric names added)

* Second pass at my Apache doc items

Changes include:

1. Breaking out the table config section from the "Running Pinot in
Production" section. I think this makes the doc easier to follow for
table configs. I have a section here which includes a sample offline and
realtime table config. I am not sure if there are some configs included
as a way for users to run a local quick start. If there is I can link to
those files in the repo, otherwise I can include the full configs on
this page for users.

2. Adding links to metrics which I am able to find in the code. I need
to collect feedback on what to do with the others - Remove or I need
pointers to where I can find them in the code.

* Updating metrics with names and URLs

The is two items I need before this is completed

1. I notice that the controller metrics appear to be generated dynamically
due to table names being present in most of the metrics. I need some advice
from the devs on how I should note these metric names so I can link to them
in the code. I want to make sure it is clear to the users.

2. I want to add sample table configs in the "in_production.rst"
section. I feel this would be best to be added in the git repo since
they will be useful for the users in the quick start guides. I need Jack
to tell me where in the repo I should put these files. Once they are in
the repo I will link to them from that section.

* Updates based on Dino's feedback

General cleanup and re-formatting based on feedback for my PR.

* A few minor changes

* Removing line numbers from metrics URLs

* Updating documentation based on PR feedback

* Modifications based on PR feedback

* Updating based on PR feedback

* Adding required table config sections

* Addressing PR comments

* Removing some un-needed headers

* Updating with more feedback from Subbu