Fixed misleading log message
It is possible that some resources converge between calls to get ServiceStatus after we
return with a minimum threshold percentage of resources good. So, we need to be careful
about not logging a misleading message about some resources not converging. We take care
to log a small number of resources that may be off, so that we don't flood the logs.