Add more logs to ServiceStatus (#4037)
* Add more logs to ServiceStatus

When we declare a service good due to its reaching the percentage threshold, it is useful to
log information about at least one partition in the resources that are not started up
correctly.

Also included a max number of resources that we log so that we don't get extra long log lines

* Addressed review comments

* Fixed misleading log message

It is possible that some resources converge between calls to get ServiceStatus after we
return with a minimum threshold percentage of resources good. So, we need to be careful
about not logging a misleading message about some resources not converging. We take care
to log a small number of resources that may be off, so that we don't flood the logs.