Fix the race condition for real-time inverted index reader (#4051)
When getting document ids from RealtimeInvertedIndexReader, the given dictionary id might not be added to the inverted index yet. We first add the value to the dictionary. Before the value is added to the inverted index, the query might have predicates that match the newly added value. In that case, the given dictionary id does not exist in the inverted index, and we return an empty bitmap.

Simplify BitmapBasedFilterOperator because getDocIds() never returns null.
Add a test for the change.