Fix 500 issue with 'GET /tenants' API call
When a server has a wrong formatted tag, current implemantation
would return 500 error for GET /tenants call.
In that case, we should not fail and return the list of valid
tenants.

For the wrong formatted tags, this pr adds the log so that one can
look up the logs when cleaning up the tags.