When refreshing the segment with identical segment, still update creation time and refresh time in ZK metadata (#4112)
We check crc to identify whether two segments are identical.
Segment creation time is not included in crc, so the segment
creation time might be different. This make sense since we might
generate segment with same data but at different time, and we don't
need to refresh the segment. In such case we should still update
creation time and refresh time.