Refactor HelixExternalViewBasedTimeBoundaryService to support all time units (#4156)
Currently we pick the segment end time as the time boundary, and
append filter 'timeColumn < boundary' to offline table and filter
'timeColumn >= boundary' to realtime table to achieve the hybrid
table federation. The problem with this is that, if the time unit
is not DAYS (for example, MILLISECONDS), and the offline table has
multiple daily segments to push, then we might get incomplete
result before all offline segments are pushed.

The solution is: always use (end time - 1 HOUR/DAY) as the time
boundary, append filter 'timeColumn <= boundary' to offline table
and 'timeColumn > boundary' to realtime table. This can ensure
all HOURLY and DAILY pushed segments be covered regardless of the
time unit, but also use as much offline side data as possible.

For HOURLY push table with time unit other than DAYS, use
(maxTimeValue - 1 HOUR) as the time boundary; otherwise, use
(maxTimeValue - 1 DAY)

Also, add checks for the time consistency between table config
and schema. We should use the time spec in schema as the source of
truth for time column because data is generated based on the
schema. In the future we might remove the timeType field from the
SegmentsValidationAndRetentionConfig.