SoC - Separate out Tuning from Translator (#4250)
Currently tuning is tied together with the translator. Each time we need to tune the detection, we have to call the translator which tries to re-convert the yaml config to detection object. This PR separates the tuner from the translator so that it can be triggered directly on the detection object.

Changes:
* Introduced ConfigTranslator interface for detection and subscription
* Enforce validation of raw and parsed config through interface (template-pattern)
* Moved out tuning related code under DetectionConfigTuner
* Move translator specific code under translator package
* Remove YamlTranslationResult; Cleaned up buildMetricUrn; Consistently use componentKey throughout instead of componentName
* Added and updated unit tests