Add RealtimeConsumptionCatchupServiceCallback (#4218)
Creating a ServiceCallback within ServiceStatus, to check that realtime consuming segments have caught up. This will help with latency issues which can happen during server startup, because the server has to catchup on a lot of consumption.
For the initial version, this only adds a static wait time (configurable). More smarts can be added later (monitor stability or idleness of stream consumption rate)