Add support for passing headers in pinot client (#4243)
* Add support for passing headers in pinot client

There are usecases where the service using the pinot client need to send the user defined headers to the http endpoint which seats behind load balancer. Adding support for the same in the client

* Add support for passing headers in pinot client

There are usecases where the service using the pinot client need to send the user defined headers to the http endpoint which seats behind load balancer. Adding support for the same in the client

* Add annotation for interface and default inplementation for new interface method

* remove unused includes

* Keep PinotClientTransportFactory interface annotation as Private

Currently this interface is not exposed as public

* Annotate ConnectionFactory appropriately which is exposed outside of the library

* Add pinot-common as dependency and use annotations of it

* Revert "Add pinot-common as dependency and use annotations of it"

This reverts commit 223e4ad31092d2e7a28cbe7380bdaf64e14ee95f.

* Remove annotations for interface from pinot-api module