Remove configurations in favor of overridden methods for test.
We seem to be introducing extra configurations just for testing purpose.
Removed the extra configs for initial delay in favor of overridden methods
for tests.

Retained the configuration for frequency, since an admin may want to
set different frequencies for the different tasks depending on the
cluster. In real clusters, the initial delay is a random one subject
to a minimum