Added tests for RealtimeSegmentValidationManager (#4306)
* Added tests for RealtimeSegmentValidationManager

Removed the dropping of realtime table and setting it up again. Relying on
the realtime table being created ahead of time

* Remove configurations in favor of overridden methods for test.

We seem to be introducing extra configurations just for testing purpose.
Removed the extra configs for initial delay in favor of overridden methods
for tests.

Retained the configuration for frequency, since an admin may want to
set different frequencies for the different tasks depending on the
cluster. In real clusters, the initial delay is a random one subject
to a minimum