Issue #4450: Support parsing of `Order By` clause in PQL compiler.
This is the first of several PR's to come for implementing group-by +
order-by support in PQL. The changes in this PR are limited to parsing
the order-by clause.

1. Added a top level orderBy field in BrokerRequest. The existing
selectionSort field is tied to selection only queries, and removing it
will break backward compatibility. Once the code is moved to use
PinotQuery instead of BrokerRequest, this will get cleaned up
automatically.

2. Moved to latest Thrift version 0.12.0. This required removing the
org.apache.commons.lang3 exclusion from pom.xml.

3. Added unit test for order-by compilation.