[TE] anomaly filters potential inconsistency issue fix (#4448)
Currently, the anomaly filters will re-fetch data from the data source to get the current value for anomalies. This works fine in most cases. However, when the data inconsistency issue happens, it might filter based on the latest fetched data instead of filtering on the current value stored in the anomaly. When this happens, it will cause confusions in the Preview UI that the filter is not working correctly. This PR fixes this issue.