Reduce the number of calls to ZK to fetch the current ideal state in TableRebalancer. Even if after updating the ideal statte (after moving segments), if we successfully persisted this in ZK (no version error), we were still going back and fetching the ideal state from ZK again.
The changes here fetch the ideal state from ZK and re-compute target
ideal state only if we failed to persist updated ideal state in ZK earlier
dut to version error. If there is no version error, we simply start the
next iteration (if there are more segments to be moved to get to target)
by taking ideal state we just wrote in ZK as the current ideal state.