Set the consumption end time of a segment based on its creation time (#4546)
We used to set the end time of a consuming segment as the consumption
start time plus the configured time threshold. In case of server restarts,
that increases the time threshold by an amount depending on how long
the consumption has alread happened.

We now base the end time on the segment creation time instead of
consumption start time.