[TE] fix multi-day granularity alignment issue (#4566)
Fix the following issues.

If monitoring granularity is larger than one day, it will align to the epoch boundary. This PR makes it align the daily boundaries instead. (The boundary alignment is required by the Pinot data source)

When making anomalies, it will filter the data frame range that is not in the monitoring window. This usually caused by the boundary alignment. It might filter useful anomalies. This PR removed this logic. Mergers will handle it gracefully.