Support serialized BYTES data with non-default factor
For the following approximate aggregation functions, the function allows a factor to tune the result size and accuracy:
- DistinctCountHLL - log2m for HyperLogLog
- PercentileEst - max error for QuantileDigest
- PercentileTDigest - compression for TDigest
When user provides serialized Object via BYTES, this factor is embedded inside the serialized bytes.
In that case, we should pick the one inside the serialized Object instead of using the default one.

Added SerializedBytesQueriesTest to test the behavior