Fix state change when consumption is stopped
When a server stops consumption due to any exception or error, it invokes
a controller operation to mark itself as being OFFLINE. It may happen that
this condition takes place right after a segment is marked as ONLINE,
but the consumer thread on the server is still alive and waiting for the
completion response (of KEEP/COMMIT/DISCARD, etc.)

While updating the ideal state, the controller should ensure that the
Helix state of the segment is CONSUMING and not anything else. If it
is not CONSUMING, then we don't want to error out, but just quietly report
back success to the server, and let it do its thing.