[TE] add event trigger listener for scheduling (#4724)
This commit mainly includes the logic of trigger event listener for Thirdeye and the event-driven scheduler will be sent in the next PR. TriggerEventListener one of the options when we run ThirdEyeAnomalyApplication and it will listen to a Kafka topic once it is enabled. Please note that this PR only defines the required interface of Kafka consumer and does not include any implementation of Kafka consumer, since Kafka event can be application-specific.