Enhance RoutingTable with ServerInstance
Currently, RoutingTable is represented as Map<String, List>, a map from serverInstanceName to list of segments hosted by the server. The rest of the code derives the hostname and port by parsing serverInstanceName assuming it's of the format: Server_<hostname>_<port>.
This PR replaces serverInstanceName with a concrete ServerInstance Object initiated from the InstanceConfig.
This will allow user to de-couple the instance name from hostname and port for the instance.

Pre-cursoe to #4525
Based on #4817
Replace #4795