Bug Fix: Fix segment download url in SegmentZkMetadata. (#4825)
It turns out that for the case when segment data directory is
deep-storage, but the segment upload method is with segment payload
being puhsed (as opposed to deep-store URI), the code had a bug where it
was generating incorrect download URL in the SegmentZkMetadata.

This PR fixes the issue with minimal change, by making the download URL
the same as the final location of the segment in deep-store. Having said
that, this part of the code seems quite complex with too many variable
controls (segment upload type, segment moving or not to final location,
etc), it would be good to clean this up at some point.