Split BrokerReduceService code into several DataTableReducers (#4851)
Splitting the code which sets results into BrokerResponseNative, in BrokerReduceService, into multiple DataTableReducers. The work in this PR is mostly copy pasting the methods from BrokerReduceService into the right Reducer, without any logic change. The only logic added, is the handling of empty dataTableMap into each setter. It was done upfront in the earlier implementation, before selecting the path based on the query.