After pulling in the latest changes, mvn clean install fails. (#4887)
The reason is that source in pinot-record-readers module is not
immediately in the sub-directory. We have 6 additional sub-modules
pinot-avro, pinot-json etc and they were not able to reference the
parent pom -- since the parent pom is not in the immediate parent
directory. relativePath by default uses "..".

The fix is to add a parent pom