Refactor file format based RecordReader constructors (#4888)
* Move all file format based record reader to be constructed using class name and init method

* After pulling in the latest changes, mvn clean install fails. (#4887)

The reason is that source in pinot-record-readers module is not
immediately in the sub-directory. We have 6 additional sub-modules
pinot-avro, pinot-json etc and they were not able to reference the
parent pom -- since the parent pom is not in the immediate parent
directory. relativePath by default uses "..".

The fix is to add a parent pom

* Move RecordReaderFactory and FileFormat to spi

* Results in ResultTable if responseFormat=sql (#4877)

* Address comments

* Move all file format based record reader to be constructed using class name and init method

* Move RecordReaderFactory and FileFormat to spi

* Address comments

* Address comments