Moving dependencies and enabling a shaded jar
Removed pinot-core dependency on pinot-thrift, pinot-json and pinot-csv.
Instead, added these as dependencies inside pinot-hadoop and pinot-spark.

Enabled shaded build for the kafka common code so that we can export the
class AvroRecordToPinotRowGenerator for internal builds.

A better way to do this may be to move check-in external kafka connectors