Fix the issue of server not registering state model factory before connecting the Helix manager
State model factory should be registered before connecting the Helix manager, while Helix property store is not available before connecting the Helix manager.
In order to solve this problem, passing Helix manager to the server instance and starting the server instance as a pre-connect callback in Helix manager.
Removed ServerBuilder to simplify the logic.