Enhance TableRebalancer to support reassigning COMPLETED segments when instances changed
Motivation:
Currently when we change the instances for a LLC realtime table,
there is no way to reassign COMPLETED segments to align with the
partitions of the CONSUMING segments. This feature is especially
important when partition-aware routing is enabled for the table.
Right now we manually calculate the assignment and replace it in
ZK, which is both troublesome and risky, and we cannot leverage
the no-downtime rebalance.

Change:
When COMPLETED instance partitions are not provided (do not
relocate COMPLETED segments), reassign COMPLETED segments the same
way as CONSUMING segments so that partitions can be aligned.
Enhance TableRebalancer to support aligning the partitions for
COMPLETED segments without downtime.