Decouple server instance id with hostname/port config. (#4995)
Currently Helix will derived hostname/port from the server instanceId string (as demonstrated in the test _testDisableLogicalServerID()_) if the instance id is set in server config. This PR decouples the configuration of server instance id and hostname/port.  This allows Pinot user to set the following field independently:
- Instance ID
- server_netty_host
- server_netty_port

This is done by overwriting the Helix zk data after server connects to the zk cluster. It is hidden behind a flag now so the current use cases will not be affected.

Also added an integration test to test different configuration choices for servers.