Change SqlConformanceLevel from LENIENT to BABEL (#5031)
The goal of doing this is to make it easier to get rid of PqlParser, and make CalciteParser the default. Pql parser was very liberal with reserved sql keywords.
This allows most of the keywords from the SqlAbstractParserImpl.getSql92ReservedWords() reserved keywords set in the query (e.g. count, max. language, position, return, module, groups).
Some exceptions to this are time, date, timestamp. table, group, which still need to be escaped.
All other syntactical validations still hold.