[TE] frontend - harleyjj/alert-details - bug fix and consolidation of anomaly filtering to a method (#5048)
The bug fix is for protecting against the anomaly.properties.detectorComponentName not being in the anomaly response, which is the case for user-reported anomalies. The code was present in the filtering logic for anomaliesOld but not anomaliesNew. The logic is duplicate and thus consolidated into a method to prevent future bug fixes from being applied to only one set of anomalies, where it should be applied to both. See key line below:

return ((anomaly.properties.detectorComponentName || '').includes(selectedRule.detectorName));