[TE] [data-availability] Scheduling of data availability tasks & some clean up (#5056)
This PR is part of the effort to enable data availability (SLA) monitoring across metrics/datasets. This particular PR only deals with the scheduling of data availability tasks.

It is also worth mentioning that we are dealing with the data availability checks only for detection jobs. In the next phase, we will have a DataSLACronScheduler which will allow users to define custom SLAs independent of detection.

Scheduling Criteria - we schedule a data availability (SLA checker) task in the following cases.
a. Cron based scheduler - Schedule a DA tasks with every detection task.
b. Event based scheduler - Schedule a DA task whenever we perform a fallback & schedule a detection.