Add EarlyTerminationException to prevent logging too many invalid error messages
When a query times out or is already satisfied, we will cancal the operator execution for early termination.
In such case, we should not log error messages as the interruptions are intentional.
In case of timeout, it will still be logged in the combine operators once, but not for each individual operators.