Support order-by on BYTES column
In order to support order-by on BYTES column everywhere, inside the system we should always use ByteArray (comparable) to store the BYTES value.
Currently BYTES value are stored as byte[], ByteArray or String in different places, which is very confusing and could cause casting errors.

Changes:
- For DisctinctCount, fix the casting issue when ordering on BYTES column
- For selection order-by, order BYTES column using ByteArray instead of String for performance improvement
- Inside Record, always store BYTES as ByteArray for clarity and also performance improvement (avoid expensive deepEquals and deepHashCode)
- On broker side, store BYTES column using ByteArray instead of String for performance improvement
- On broker side, support type compatible merges for all selection queries

No format change on the query results.
TODO: We are still returning String for BYTES column when preserving the type. Consider changing it to byte[].