[TE] Add validations for the threshold rule anomaly filter
The threshold rule anomaly filter is used for filtering the anomaly based on the current value in the anomaly duration.
If the metric is not aggregated by SUM or COUNT, user may set the wrong parameter and cause the system to filter
the anomaly based on a wrong value. This PR fixes the issue.