Disable S3MockListner in S3PinotFSTest. (#5292)
We have been running into build failures with this test. Disabling the
test did not help as the failure was actually coming from the S3MockListner
that tries to start S3Mock (and fails with port already in use). Given that
there is no api to choose port to start S3Mock at, we are temporarily disabling
the test and unregistering this listener.