In SegmentPurger, use table config to generate the segment
All the indexing and partition info are stored in the table config.
Without the table config, purged segment will lose the indexing and partition info, which could cause performance issue.

In the following pr, we will enforce all segment creation to provide table config and schema to guarantee the consistent behavior.