Timespec to datetimespec conversion utility (#5324)
This is an item of project: #2756
We plan to internally start treating timeFieldSpec as dateTimeFieldSpec. This PR adds a utility function which converts a timeFieldSpec to an equivalent dateTimeFieldSpec.
Note that dateTimeFieldSpec doesn't have the concept of incoming/outgoing, and hence we construct and add a transform function to convey the conversion. Introduced a lot of epoch time transform functions for this.

NOTE: the toEpochXXXBucket method is added to help with conversion from timeFieldSpec to dateTimeFieldSpec. Practically, we would only be using toEpochXXX and toEpochXXXRounded.