Optimize RealtimeDictionaryBasedRangePredicateEvaluator by not scanning the dictionary when cardinality is high
For real-time range predicate, because the dictionary is not sorted, in order to get the matching dictionary ids, we have to scan the whole dictionary.
This will cause performance issue when the cardinality is high for the column.
Optimize it by adding a cardinality threshold (1000 for now) to decide whether to pre-calculate all the matching dictionary ids.