Fix new broker to old server backward compatibility issue.
With PR #5259 and #5275, the broker starts to send aggregation function arguments
in a new field in the thrift class. While new server prefers new field and falls back
to old field, in case of new broker and old server the server is unable to find
values in the field.

This PR fixes this issue by adding both old and new field in the broker. We will need to
change back broker to stop setting the old field with a future release.