Remove excessive SchemaBuilder#addTime methods (#5349)
When timeFieldSpec gets deprecated in favor of DateTimeFieldSpec, one of the changes will be in the SchemaBuilder class. SchemaBuilder is used only in tests (and some external integrations maybe).
SchemaBuilder has ~10 methods to addTime. These will be replaced with addDateTime (barring a few places where we intentionally want to add TIME to test that we get DATE_TIME.) Removing most of the addTime methods, keeping only one