Introduced StreamPartitionMsgOffset class (#5360)
This change is mostly mechanical to use the new offset class (instead of a long offset)
on the server side. Segment completion protocol and segment zk metadata still hold the
offset in long format.

The StreamPartitionMsgOffset should eventually become an interface. For now, we just
hold a long value there.