Temporary work-around to bypass function arguments being treated as expressions with columns.
Only aggregation functions have the knowledge on how to interpret their arguments.
However, the assumption within ServerQueryRequest is that all expression arguments will contain
columns to be collected. A clean fix would be to create aggregation functions upfront and let
them interpret the arguments. However, until that happens, working-around by assuming that first
argument contains all the columns (to be used by DataSchemaPruner), which is true for all functions.

A side effect would be that DataSchemaPruner won't work for DistinctCountThetaSketchAggregatinoFunction.
But since this pruner is only for schema mis-matches, and not for performance, there should not be
an impact.